+++ include/clap/events.h
@@ -54,8 +54,8 @@ enum {
    // cases:
    // - a plugin is inside a drum pad in Bitwig Studio's drum machine, and this pad is choked by
    //   another one
-   // - the user double clicks the DAW's stop button in the transport which then stops the sound on
-   //   every tracks
+   // - the user double-clicks the DAW's stop button in the transport which then stops the sound on
+   //   every track
    //
    // NOTE_END is sent by the plugin to the host. The port, channel, key and note_id are those given
    // by the host in the NOTE_ON event. In other words, this event is matched against the
@@ -221,7 +221,7 @@ typedef struct clap_event_transport {
    clap_sectime  song_pos_seconds; // position in seconds
 
    double tempo;     // in bpm
-   double tempo_inc; // tempo increment for each samples and until the next
+   double tempo_inc; // tempo increment for each sample and until the next
                      // time info event
 
    clap_beattime loop_start_beats;
@@ -260,7 +260,7 @@ typedef struct clap_event_midi2 {
    uint32_t data[4];
 } clap_event_midi2_t;
 
-// Input event list, events must be sorted by time.
+// Input event list. The host will deliver these sorted in sample order.
 typedef struct clap_input_events {
    void *ctx; // reserved pointer for the list
 
@@ -271,7 +271,7 @@ typedef struct clap_input_events {
    const clap_event_header_t *(CLAP_ABI *get)(const struct clap_input_events *list, uint32_t index);
 } clap_input_events_t;
 
-// Output event list, events must be sorted by time.
+// Output event list. The plugin must insert events in sample sorted order when inserting events
 typedef struct clap_output_events {
    void *ctx; // reserved pointer for the list