+++ include/clap/ext/params.h
@@ -21,7 +21,7 @@
 ///
 /// When the plugin changes a parameter value, it must inform the host.
 /// It will send @ref CLAP_EVENT_PARAM_VALUE event during process() or flush().
-/// If the user is adjusting the value, don't forget to mark the begining and end
+/// If the user is adjusting the value, don't forget to mark the beginning and end
 /// of the gesture by sending CLAP_EVENT_PARAM_GESTURE_BEGIN and CLAP_EVENT_PARAM_GESTURE_END
 /// events.
 ///
@@ -55,8 +55,9 @@
 /// - the plugin is responsible for sending the parameter value to its audio processor
 /// - call clap_host_params->request_flush() or clap_host->request_process().
 /// - when the host calls either clap_plugin->process() or clap_plugin_params->flush(),
-///   send an automation event and don't forget to set begin_adjust,
-///   end_adjust and should_record flags
+///   send an automation event and don't forget to wrap the parameter change(s)
+///   with CLAP_EVENT_PARAM_GESTURE_BEGIN and CLAP_EVENT_PARAM_GESTURE_END to define the
+///   beginning and end of the gesture.
 ///
 /// IV. Turning a knob via automation
 /// - host sends an automation point during clap_plugin->process() or clap_plugin_params->flush().
@@ -74,7 +75,7 @@
 ///     call clap_host_params.clear(host, param_id, CLAP_PARAM_CLEAR_ALL)
 ///   - call clap_host_params->rescan(CLAP_PARAM_RESCAN_ALL)
 ///
-/// CLAP allows the plugin to change the parameter range, yet the plugin developper
+/// CLAP allows the plugin to change the parameter range, yet the plugin developer
 /// should be aware that doing so isn't without risk, especially if you made the
 /// promise to never change the sound. If you want to be 100% certain that the
 /// sound will not change with all host, then simply never change the range.
@@ -89,7 +90,7 @@
 /// should be stored as plain value in the document.
 ///
 /// If the host goes with the first approach, there will still be situation where the
-/// sound may innevitably change. For example, if the plugin increase the range, there
+/// sound may inevitably change. For example, if the plugin increase the range, there
 /// is an automation playing at the max value and on top of that an LFO is applied.
 /// See the following curve:
 ///                                   .
@@ -97,15 +98,29 @@
 ///          .....                  .   .
 /// before: .     .     and after: .     .
 ///
+/// Persisting parameter values:
+///
+/// Plugins are responsible for persisting their parameter's values between
+/// sessions by implementing the state extension. Otherwise parameter value will
+/// not be recalled when reloading a project. Hosts should _not_ try to save and
+/// restore parameter values for plugins that don't implement the state
+/// extension.
+///
 /// Advice for the host:
+///
 /// - store plain values in the document (automation)
 /// - store modulation amount in plain value delta, not in percentage
 /// - when you apply a CC mapping, remember the min/max plain values so you can adjust
+/// - do not implement a parameter saving fall back for plugins that don't
+///   implement the state extension
 ///
 /// Advice for the plugin:
+///
 /// - think carefully about your parameter range when designing your DSP
 /// - avoid shrinking parameter ranges, they are very likely to change the sound
 /// - consider changing the parameter range as a tradeoff: what you improve vs what you break
+/// - make sure to implement saving and loading the parameter values using the
+///   state extension
 /// - if you plan to use adapters for other plugin formats, then you need to pay extra
 ///   attention to the adapter requirements
 
@@ -120,7 +135,7 @@ enum {
    // if so the double value is converted to integer using a cast (equivalent to trunc).
    CLAP_PARAM_IS_STEPPED = 1 << 0,
 
-   // Useful for for periodic parameters like a phase
+   // Useful for periodic parameters like a phase
    CLAP_PARAM_IS_PERIODIC = 1 << 1,
 
    // The parameter should not be shown to the user, because it is currently not used.
@@ -191,7 +206,7 @@ typedef struct clap_param_info {
    clap_param_info_flags flags;
 
    // This value is optional and set by the plugin.
-   // Its purpose is to provide a fast access to the plugin parameter object by caching its pointer.
+   // Its purpose is to provide fast access to the plugin parameter object by caching its pointer.
    // For instance:
    //
    // in clap_plugin_params.get_info():