+++ include/clap/process.h
@@ -53,11 +53,11 @@ typedef struct clap_process {
    uint32_t                   audio_inputs_count;
    uint32_t                   audio_outputs_count;
 
-   // Input and output events.
-   //
-   // Events must be sorted by time.
    // The input event list can't be modified.
+   // Input read-only event list. The host will deliver these sorted in sample order.
    const clap_input_events_t  *in_events;
+
+   // Output event list. The plugin must insert events in sample sorted order when inserting events
    const clap_output_events_t *out_events;
 } clap_process_t;